[move] 配置場所を変更
